+++ src/PCEHR/IUploadDocumentMetadataClient.cs
@@ -23,6 +23,9 @@ namespace Nehta.VendorLibrary.PCEHR
         /// <param name="healthcareFacilityTypeCode">Healthcare facility type code.</param>
         /// <param name="practiceSetting">Practice setting code.</param>
         /// <param name="uuidOfDocumentToReplace">UUID of document to replace.</param>
+        /// <param name="documentSubTypeCode">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeCodeSystem">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeName">Optional subtype code for custom subtype docs.</param>
         /// <returns>The populated request object.</returns>
         SubmitObjectsRequest CreateRequestForReplacement(
             byte[] cdaPackageContent,
@@ -31,7 +34,10 @@ namespace Nehta.VendorLibrary.PCEHR
             string formatCodeName,
             HealthcareFacilityTypeCodes healthcareFacilityTypeCode,
             PracticeSettingTypes practiceSetting,
-            string uuidOfDocumentToReplace);
+            string uuidOfDocumentToReplace,
+            string documentSubTypeCode = "",
+            string documentSubTypeCodeSystem = "",
+            string documentSubTypeName = "");
 
         /// <summary>
         /// Helper method to generate the request object when registering a new document on the PCEHR.
@@ -42,6 +48,9 @@ namespace Nehta.VendorLibrary.PCEHR
         /// <param name="formatCodeName">Format code display name.</param>
         /// <param name="healthcareFacilityTypeCode">Healthcare facility type code.</param>
         /// <param name="practiceSetting">Practice setting code.</param>
+        /// <param name="documentSubTypeCode">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeCodeSystem">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeName">Optional subtype code for custom subtype docs.</param>
         /// <returns>The populated request object.</returns>
         SubmitObjectsRequest CreateRequestForNewDocument(
             byte[] cdaPackageContent,
@@ -49,6 +58,9 @@ namespace Nehta.VendorLibrary.PCEHR
             string formatCode,
             string formatCodeName,
             HealthcareFacilityTypeCodes healthcareFacilityTypeCode,
-            PracticeSettingTypes practiceSetting);
+            PracticeSettingTypes practiceSetting,
+            string documentSubTypeCode = "",
+            string documentSubTypeCodeSystem = "",
+            string documentSubTypeName = "");
     }
 }