+++ src/PCEHR/UploadDocumentMetadataClient.cs
@@ -110,6 +110,9 @@ namespace Nehta.VendorLibrary.PCEHR
         /// <param name="healthcareFacilityTypeCode">Healthcare facility type code.</param>
         /// <param name="practiceSetting">Practice setting code.</param>
         /// <param name="uuidOfDocumentToReplace">UUID of document to replace.</param>
+        /// <param name="documentSubTypeCode">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeCodeSystem">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeName">Optional subtype code for custom subtype docs.</param>
         /// <returns>The populated request object.</returns>
         public SubmitObjectsRequest CreateRequestForReplacement(
             byte[] cdaPackageContent,
@@ -118,7 +121,10 @@ namespace Nehta.VendorLibrary.PCEHR
             string formatCodeName,
             HealthcareFacilityTypeCodes healthcareFacilityTypeCode,
             PracticeSettingTypes practiceSetting,
-            string uuidOfDocumentToReplace)
+            string uuidOfDocumentToReplace,
+            string documentSubTypeCode = "",
+            string documentSubTypeCodeSystem = "",
+            string documentSubTypeName = "")
         {
             Validation.ValidateArgumentRequired("uuidOfDocumentToReplace", uuidOfDocumentToReplace);
 
@@ -129,7 +135,10 @@ namespace Nehta.VendorLibrary.PCEHR
                formatCodeName,
                healthcareFacilityTypeCode,
                practiceSetting,
-               uuidOfDocumentToReplace
+               uuidOfDocumentToReplace,
+               documentSubTypeCode,
+               documentSubTypeCodeSystem,
+               documentSubTypeName
                );
         }
 
@@ -142,6 +151,9 @@ namespace Nehta.VendorLibrary.PCEHR
         /// <param name="formatCodeName">Format code display name.</param>
         /// <param name="healthcareFacilityTypeCode">Healthcare facility type code.</param>
         /// <param name="practiceSetting">Practice setting code.</param>
+        /// <param name="documentSubTypeCode">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeCodeSystem">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeName">Optional subtype code for custom subtype docs.</param>
         /// <returns>The populated request object.</returns>
         public SubmitObjectsRequest CreateRequestForNewDocument(
             byte[] cdaPackageContent,
@@ -149,7 +161,10 @@ namespace Nehta.VendorLibrary.PCEHR
             string formatCode,
             string formatCodeName,
             HealthcareFacilityTypeCodes healthcareFacilityTypeCode,
-            PracticeSettingTypes practiceSetting)
+            PracticeSettingTypes practiceSetting,
+            string documentSubTypeCode = "",
+            string documentSubTypeCodeSystem = "",
+            string documentSubTypeName = "")
         {
             return CreateRequest(
                 cdaPackageContent,
@@ -158,7 +173,10 @@ namespace Nehta.VendorLibrary.PCEHR
                 formatCodeName,
                 healthcareFacilityTypeCode,
                 practiceSetting,
-                null
+                null,
+                documentSubTypeCode,
+                documentSubTypeCodeSystem,
+                documentSubTypeName
                 );
         }
 
@@ -172,7 +190,10 @@ namespace Nehta.VendorLibrary.PCEHR
         /// <param name="healthcareFacilityTypeCode">Healthcare facility type code.</param>
         /// <param name="practiceSetting">Practice setting code.</param>
         /// <param name="uuidOfDocumentToReplace">UUID of document to replace. NULL for new documents.</param>
-        /// <returns>The populated request object.</returns>
+        /// <param name="documentSubTypeCode">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeCodeSystem">Optional subtype code for custom subtype docs.</param>
+        /// <param name="documentSubTypeName">Optional subtype code for custom subtype docs.</param>
+        /// /// <returns>The populated request object.</returns>
         internal SubmitObjectsRequest CreateRequest(
             byte[] cdaPackageContent,
             string uniqueRepositoryId,
@@ -180,7 +201,10 @@ namespace Nehta.VendorLibrary.PCEHR
             string formatCodeName,
             HealthcareFacilityTypeCodes healthcareFacilityTypeCode,
             PracticeSettingTypes practiceSetting,
-            string uuidOfDocumentToReplace)
+            string uuidOfDocumentToReplace,
+            string documentSubTypeCode = "",
+            string documentSubTypeCodeSystem = "",
+            string documentSubTypeName = "")
         {
             Validation.ValidateArgumentRequired("cdaPackageContent", cdaPackageContent);
             Validation.ValidateArgumentRequired("uniqueRepositoryId", uniqueRepositoryId);
@@ -201,7 +225,10 @@ namespace Nehta.VendorLibrary.PCEHR
                 cdaPackageContent.Length,
                 CalculateSHA1(cdaPackageContent),
                 true,
-                uuidOfDocumentToReplace
+                uuidOfDocumentToReplace,
+                documentSubTypeCode,
+                documentSubTypeCodeSystem,
+                documentSubTypeName
                 );
 
             var sor = metadata.CreateSubmitObjectsRequest();